Created Godot Folder
Here's the folder with our project in it. Right now, it just had a main scene with nothing in it.

Probably best plan is to work on scenes separately, and then instance them into the main scene when we're done so that Git doesn't have to freak out trying to figure out what changed.